pyproject.toml
[project]
name = "dessaux-damien-ecf1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fake-useragent>=2.2.0",
    "lxml>=6.0.2",
    "minio>=7.2.20",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
]
